Backend/requirements.txt
Django==5.2.7
djangorestframework
Pillow
whitenoise
psycopg2-binary
PyMySQL
Faker
django-cors-headers
python-dotenv
gunicorn
django-unfold